move it all to a separate processor
