+++ go.work
@@ -5,4 +5,5 @@ toolchain go1.24.3
 use (
 	./otelcol-dev
 	./symbolicatorprocessor
+	./dsymprocessor
 )
