+++ symbolicatorprocessor/config.go
@@ -34,14 +34,6 @@ type Config struct {
 	// StackMessageKey is the attribute key that contains the message of the stack trace.
 	StackMessageKey string `mapstructure:"stack_message_key"`
 
-	// MetricKitStackTraceAttributeKey is the attribute key that contains the metrickit
-	// stack trace.
-	MetricKitStackTraceAttributeKey string `mapstructure:"metrickit_stack_trace_attribute_key"`
-
-	// OutputMetricKitStackTraceAttributeKey is the attribute key that contains the
-	// symbolicated metrickit stack trace.
-	OutputMetricKitStackTraceAttributeKey string `mapstructure:"output_metrickit_stack_trace_attribute_key"`
-
 	// preserveStackTrace is a config option that determines whether to keep the
 	// original stack trace in the output.
 	PreserveStackTrace bool `mapstructure:"preserve_stack_trace"`
@@ -81,9 +73,6 @@ type Config struct {
 
 	// CacheSize is the maximum number of source maps to cache.
 	SourceMapCacheSize int `mapstructure:"source_map_cache_size"`
-
-	// CacheSize is the maximum number of dSYMs to cache.
-	DSYMCacheSize int `mapstructure:"dsym_cache_size"`
 }
 
 type LocalSourceMapConfiguration struct {
